Rename generator to customends
