+++ internal/generate/customends/custom_endpoints_footer.tmpl
