+++ internal/generate/customends/custom_endpoints_header.tmpl
