+++ internal/generate/customends/generate.go
@@ -1,4 +1,4 @@
 //go:generate go run main.go
 // ONLY generate directives and package declaration! Do not add anything else to this file.
 
-package website
+package customends
