+++ internal/generate/customends/main.go
